requirements.txt
streamlit
pydub
SpeechRecognition
arabic-reshaper
python-bidi
pyaudio
